コメント削除 / 型宣言をさらに分離
