+++ server/db/general.go
@@ -5,6 +5,7 @@ import (
 	"net"
 	"os"
 	"time"
+	"traQ-gazer/model"
 
 	"github.com/go-sql-driver/mysql"
 	"github.com/jmoiron/sqlx"
@@ -24,10 +25,6 @@ var (
 )
 
 func SetUp() error {
-	// jst, err := time.LoadLocation("Asia/Tokyo")
-	// if err != nil {
-	// 	return err
-	// }
 	conf := mysql.Config{
 		User:                 db_username,
 		Passwd:               db_password,
@@ -69,9 +66,9 @@ func initUsersTable() error {
 		return err
 	}
 
-	userList := UserList{}
+	userList := model.UserList{}
 	for _, user := range result {
-		userList = append(userList, User{Traq_uuid: user.Id, Trap_id: user.Name, Is_bot: user.Bot})
+		userList = append(userList, model.User{Traq_uuid: user.Id, Trap_id: user.Name, Is_bot: user.Bot})
 	}
 
 	alreadyExistUsersUUIDList := []string{}
@@ -93,14 +90,6 @@ func initUsersTable() error {
 	return nil
 }
 
-type User struct {
-	Traq_uuid string `db:"traq_uuid"`
-	Trap_id   string `db:"trap_id"`
-	Is_bot    bool   `db:"is_bot"`
-}
-
-type UserList []User
-
 func min(a, b int) int {
 	if a < b {
 		return a
@@ -108,11 +97,11 @@ func min(a, b int) int {
 	return b
 }
 
-func removeAlreadyExistUsers(allUsers UserList, alreadyUsersUUID []string) UserList {
-	newUserList := make(UserList, 0)
+func removeAlreadyExistUsers(allUsers model.UserList, alreadyUsersUUID []string) model.UserList {
+	newUserList := make(model.UserList, 0)
 	for _, all := range allUsers {
 		if !slices.Contains(alreadyUsersUUID, all.Traq_uuid) {
-			newUserList = append(newUserList, User{
+			newUserList = append(newUserList, model.User{
 				Traq_uuid: all.Traq_uuid,
 				Trap_id:   all.Trap_id,
 				Is_bot:    all.Is_bot,
